cmdline: allow rplint to be called with 'python -m'
Allow rplint __main__ to be invoked with the -m flag
without raising RuntimeWarning:

	$ python3 -m rplint --help
	Usage: rplint [OPTIONS] FILENAME

	Options:
	  -l, --line-length INTEGER
	  -j, --jima                 use extra bad word list
	  -h, --help                 Show this message and exit.

This also removes a RuntimeWarning by getting __main__
out of direct import into __init__. Previous:

	$ python3 -m rplint -h
	runpy.py:127: RuntimeWarning: 'rplint.__main__' found in
	sys.modules after import of package 'rplint', but prior to
	execution of 'rplint.__main__'; this may result in
	unpredictable behaviour
	  warn(RuntimeWarning(msg))
